refactor: move product variant selection test into product variant test
This test was originally in a separate file and when renaming the
test case names and test file names I noticed an overlap. Both test files
were testing interactions with the product variants. Therefore, it makes
more sense to have them in the same file.
